feat(core/constants): add some global types and checkers

+++ src/core/api.ts
@@ -1,10 +1,11 @@
 import 'dotenv/config';
 import { ZaCoCiaranCumulusBet, ZaCoCiaranCumulusMarket, ZaCoCiaranCumulusResolution } from '../../generated/typescript';
-import { is, type ActorIdentifier } from '@atcute/lexicons';
+import { is, type ActorIdentifier, type Did } from '@atcute/lexicons';
 import type { CreateCommit, DeleteCommit } from '@atcute/jetstream';
 import { DEFAULT_MARKET_COLS, DEFAULT_BET_COLS, DEFAULT_RESOLUTION_COLS, marketsTable, betsTable, resolutionsTable } from '@/db'
 import { eq } from 'drizzle-orm';
 import { db } from '@/db';
+import { createUri, isBetCollection, isMarketCollection, isResolutionCollection } from './utils';
 
 export async function tryListMarkets() {
     return await db.query.marketsTable.findMany({
@@ -56,13 +57,16 @@ export async function tryFindMarketResolutions(uri: string) {
     })
 }
 
-export async function tryCreateMarket(did: ActorIdentifier, { record, rev, rkey, cid }: CreateCommit) {
+export async function tryCreateMarket(did: Did, { record, rev, rkey, cid }: CreateCommit) {
     if (is(ZaCoCiaranCumulusMarket.mainSchema, record)) {
-        const uri = `at://${did}/${record.$type}/${rkey}`;
+        const uri = createUri(did, record.$type, rkey);
         console.log("> Creating Market:", uri);
 
         const { question, liquidity } = record;
         const [closesAt, createdAt] = [new Date(record.closesAt), new Date(record.createdAt)];
+
+        const existing = await db.query.marketsTable.findFirst({ where: eq(marketsTable.uri, uri) });
+        if (existing) return;
 
         const marketData: typeof marketsTable.$inferInsert = {
             uri, did, rev, rkey, cid, question, liquidity, closesAt, record, createdAt,
@@ -73,18 +77,23 @@ export async function tryCreateMarket(did: ActorIdentifier, { record, rev, rkey,
     }
 }
 
-export async function tryDeleteMarket(did: ActorIdentifier, commit: DeleteCommit) {
-    const uri = `at://${did}/${commit.collection}/${commit.rkey}`;
+export async function tryDeleteMarket(did: Did, commit: DeleteCommit) {
+    if (!isMarketCollection(commit.collection)) return;
+    const uri = createUri(did, commit.collection, commit.rkey);
     console.log("> Deleting Market:", uri);
     await db.delete(marketsTable).where(eq(marketsTable.uri, uri))
 }
 
-export async function tryCreateBet(did: ActorIdentifier, { record, rev, rkey, cid }: CreateCommit) {
+export async function tryCreateBet(did: Did, { record, rev, rkey, cid }: CreateCommit) {
     if (is(ZaCoCiaranCumulusBet.mainSchema, record)) {
-        const uri = `at://${did}/${record.$type}/${rkey}`;
+        const uri = createUri(did, record.$type, rkey);
         console.log("> Creating Bet:", uri);
+
         const { position, market } = record;
         const createdAt = new Date(record.createdAt);
+
+        const existing = await db.query.betsTable.findFirst({ where: eq(betsTable.uri, uri) });
+        if (existing) return;
 
         const betData: typeof betsTable.$inferInsert = {
             uri, did, rev, rkey, cid, position, marketUri: market.uri, record, createdAt
@@ -95,18 +104,23 @@ export async function tryCreateBet(did: ActorIdentifier, { record, rev, rkey, ci
     }
 }
 
-export async function tryDeleteBet(did: ActorIdentifier, commit: DeleteCommit) {
-    const uri = `at://${did}/${commit.collection}/${commit.rkey}`;
+export async function tryDeleteBet(did: Did, commit: DeleteCommit) {
+    if (!isBetCollection(commit.collection)) return;
+    const uri = createUri(did, commit.collection, commit.rkey);
     console.log("> Deleting Bet:", uri);
     await db.delete(betsTable).where(eq(betsTable.uri, uri))
 }
 
-export async function tryCreateResolution(did: ActorIdentifier, { record, rev, rkey, cid }: CreateCommit) {
+export async function tryCreateResolution(did: Did, { record, rev, rkey, cid }: CreateCommit) {
     if (is(ZaCoCiaranCumulusResolution.mainSchema, record)) {
-        const uri = `at://${did}/${record.$type}/${rkey}`;
+        const uri = createUri(did, record.$type, rkey);
         console.log("> Creating Resolution:", uri);
+
         const { answer, market } = record;
         const createdAt = new Date(record.createdAt);
+
+        const existing = await db.query.resolutionsTable.findFirst({ where: eq(resolutionsTable.uri, uri) });
+        if (existing) return;
 
         const resolutionData: typeof resolutionsTable.$inferInsert = {
             uri, did, rev, rkey, cid, answer, marketUri: market.uri, record, createdAt
@@ -117,8 +131,9 @@ export async function tryCreateResolution(did: ActorIdentifier, { record, rev, r
     }
 }
 
-export async function tryDeleteResolution(did: ActorIdentifier, commit: DeleteCommit) {
-    const uri = `at://${did}/${commit.collection}/${commit.rkey}`;
+export async function tryDeleteResolution(did: Did, commit: DeleteCommit) {
+    if (!isResolutionCollection(commit.collection)) return;
+    const uri = createUri(did, commit.collection, commit.rkey);
     console.log("> Deleting Resolution:", uri);
     await db.delete(resolutionsTable).where(eq(resolutionsTable.uri, uri))
 }
